remove `org member get` endpoint
There's no endpoint for this in the API, only the `list` endpoint
exists in the API.  So it doesn't really make sense to duplicate that
functionality in an `org member get` command.  Just remove the code
the clanker wrote, so we can close out the issue the other clanker
opened unnecessarily.

Fixes https://github.com/pulumi/pulumi/issues/23011
